# Conflicts:
#	documents/sqs/misc/sqs-manifest
#	documents/sqs/test/queue_state_failure_dlq_fifo/2020-11-27/Documents/AutomationDocument.yml
#	documents/sqs/test/queue_state_failure_dlq_fifo/2020-11-27/Tests/features/queue_state_failure_dlq_fifo.feature
#	documents/sqs/test/queue_state_failure_dlq_fifo/2020-11-27/Tests/step_defs/test_queue_state_failure_dlq_fifo.py
#	documents/util/scripts/src/ssm_execution_util.py
#	resource_manager/cloud_formation_templates/SqsTemplate.yml

+++ resource_manager/src/resource_manager.py
@@ -31,11 +31,13 @@ class ResourceManager:
                     return rt
             raise Exception('Resource type for name [{}] is not supported.'.format(resource_type))
 
-    def __init__(self, cfn_helper: CloudFormationTemplate, s3_helper: S3):
+    def __init__(self, cfn_helper: CloudFormationTemplate, s3_helper: S3, custom_pool_size: dict, test_session_id: str):
         self.cfn_helper = cfn_helper
         self.s3_helper = s3_helper
         self.cfn_templates = dict()
         self.cfn_resources = dict()
+        self.custom_pool_size = custom_pool_size
+        self.test_session_id = test_session_id
 
     def init_ddb_tables(self, boto3_session):
         """
@@ -158,6 +160,7 @@ class ResourceManager:
                                 resource.status = ResourceModel.Status.LEASED.name
                                 resource.leased_on = datetime.now()
                                 resource.updated_on = datetime.now()
+                                resource.test_session_id = self.test_session_id
                                 resource.save()
                                 return resource
                 except PutError:
@@ -207,15 +210,15 @@ class ResourceManager:
         """
         logging.info("Releasing all stalled resources.")
         for resource in ResourceModel().scan():
-            # If resource was not released because of failure or cancellation
-            if resource.status == ResourceModel.Status.LEASED.name:
-                ResourceModel.update_resource_status(resource, ResourceModel.Status.AVAILABLE)
-
-            # If resource was not fully created/updated because of failure or cancellation
-            elif resource.status != ResourceModel.Status.AVAILABLE.name:
-                logging.info('Deleting resource for stack name [{}] in status [{}].'.format(resource.cf_stack_name,
-                                                                                            resource.status))
-                resource.delete()
+            if not self.test_session_id or resource.test_session_id == self.test_session_id:
+                # If resource was not released because of failure or cancellation
+                if resource.status == ResourceModel.Status.LEASED.name:
+                    ResourceModel.update_resource_status(resource, ResourceModel.Status.AVAILABLE)
+                # If resource was not fully created/updated because of failure or cancellation
+                elif resource.status != ResourceModel.Status.AVAILABLE.name:
+                    logging.info('Deleting resource for stack name [{}] in status [{}].'.format(resource.cf_stack_name,
+                                                                                                resource.status))
+                    resource.delete()
 
     def destroy_all_resources(self):
         """
@@ -252,21 +255,26 @@ class ResourceManager:
         logging.info("Destroying [%s] stack.", resource.cf_stack_name)
         cfn_helper.delete_cf_stack(resource.cf_stack_name)
 
-    def _get_resource_pool_size(self, cf_template_name, resource_type: ResourceType):
+    def _get_resource_pool_size(self, cfn_template_name: str, resource_type: ResourceType) -> int:
         """
-        Using config.py finds pool size for given template
-        (number of cloud formation stack copies to be created for given cloud formation template).
-        :param cf_template_name The file path to cloud formation template
+        Finds pool size (number of cloud formation stack copies to be created for given cloud formation template)
+        defined in config.pool_size or custom_pool_size for given template
+        if custom_pool_size has configuration for cfn_template_name it will override config.pool_size.
+        :param cfn_template_name The file path to cloud formation template
         :return The pool size
         """
-        pool_size = 1 if resource_type == ResourceManager.ResourceType.ASSUME_ROLE \
-            else config.pool_size.get(cf_template_name)
-        if pool_size is None:
-            default_pool_size = config.pool_size['default']
-            logging.warning("Pool size for [%s] template not found, using default: %d",
-                            cf_template_name, default_pool_size)
-            return default_pool_size
-        logging.info("Pool size for [%s] template: %d", cf_template_name, pool_size)
+        pool_size = 1
+        if resource_type == ResourceManager.ResourceType.ASSUME_ROLE:
+            return pool_size
+        elif self.custom_pool_size and self.custom_pool_size.get(cfn_template_name):
+            pool_size = self.custom_pool_size.get(cfn_template_name)
+        else:
+            pool_size = config.pool_size.get(cfn_template_name)
+            if pool_size is None:
+                pool_size = config.pool_size['default']
+                logging.warning("Pool size for [%s] template not found, using default: %d",
+                                cfn_template_name, pool_size)
+        logging.info("Pool size for [%s] template: %d", cfn_template_name, pool_size)
         return pool_size
 
     def _update_resource(self, index: int, cfn_template_path: str, cfn_content: dict, resource: ResourceModel):
@@ -286,6 +294,7 @@ class ResourceManager:
 
             # Changing status to UPDATING to block other threads to use it.
             resource.status = ResourceModel.Status.UPDATING.name
+            resource.test_session_id = self.test_session_id
             resource.save()
 
             # Updating cloud formation stack stack
@@ -328,7 +337,8 @@ class ResourceManager:
                 status=ResourceModel.Status.CREATING.name,
                 leased_on=datetime.now(),
                 created_on=datetime.now(),
-                updated_on=datetime.now()
+                updated_on=datetime.now(),
+                test_session_id=self.test_session_id
             )
 
             # Creating cloud formation stack stack
@@ -367,6 +377,7 @@ class ResourceManager:
         resource.leased_times = resource.leased_times + 1
         resource.cf_input_parameters = cfn_input_params
         resource.cf_output_parameters = cf_output_params
+        resource.test_session_id = self.test_session_id
         resource.save()
         return resource
